Move the HA device functions to separate files
homeassistant_discovery.cpp was becoming quite large and not even all
services are added yet. So before it becomes unmanageble, the virtual
and support functions of the separate devices are moved to a file per
device type.
